Updated README

+++ app.py
@@ -9,19 +9,18 @@ from flask_cors import CORS
 from config import Config
 from models import db, Project, Pathogen, Study
 import jwt
+from jwt.exceptions import ExpiredSignatureError, InvalidTokenError
 import datetime
+import requests
+
 
 app = Flask(__name__)
 app.config.from_object(Config)
 CORS(app, resources={r"/api/*": {"origins": "*", "methods": ["GET", "POST", "PUT", "DELETE"]}})
 
-if not hasattr(app, 'json_encoder'):
-    app.json_encoder = JSONEncoder
-
-
 db.init_app(app)
 migrate = Migrate(app, db)
-jwtm = JWTManager(app)
+
 api = SwaggerApi(app, api_version='0.1', api_spec_url='/swagger')
 
 SWAGGER_URL = '/swagger'
@@ -35,14 +34,68 @@ swaggerui_blueprint = get_swaggerui_blueprint(
 )
 app.register_blueprint(swaggerui_blueprint, url_prefix=SWAGGER_URL)
 
+if not hasattr(app, 'json_encoder'):
+    app.json_encoder = JSONEncoder
+
+def get_public_key():
+    try:
+        response = requests.get(app.config['PUBLIC_KEY_URL'])
+        response.raise_for_status()  
+        public_key = response.text
+        return public_key
+    except requests.exceptions.RequestException as e:
+        print(f"Error fetching public key: {e}")
+        return None
+
+app.config['JWT_PUBLIC_KEY'] = get_public_key()
+
+jwtm = JWTManager(app)
+
+# Proxy to SONG API. Fix this.
+@app.route('/api/schemas/', methods=['GET'])
+def get_schemas():
+    try:
+        response = requests.get(app.config['SONG_API'] + '/schemas',)
+        response.raise_for_status()
+        schemas = response.json()
+        return schemas
+    except requests.exceptions.RequestException as e:
+        print(f"Error fetching schemas: {e}")
+        return None
+
+
+    
+
 class User(Resource):
-    @jwt_required()
     def get(self):
-        claims = get_jwt() 
-        return claims
+        auth_header = request.headers.get('Authorization')
+
+        if not auth_header:
+            return {"error": "Authorization header missing"}, 401
+
+        public_key = get_public_key()
+
+        parts = auth_header.split()
+
+        if len(parts) != 2 or parts[0] != 'Bearer':
+            return {"error": "Invalid Authorization header format"}, 401
+
+        token = parts[1]
+
+        try:
+            payload = jwt.decode(token, public_key, algorithms=['RS256'])
+            return payload, 200
+        except ExpiredSignatureError:
+            print('Expired token')
+            return {"error": "Expired token"}, 401
+        except InvalidTokenError:
+            print('Invalid token')
+            return {"error": "Invalid token"}, 401
+
+       
+
 
 class Projects(Resource):
-
     @swagger.doc({
         'tags': ['projects'],
         'description': 'Get all projects',
@@ -89,6 +142,13 @@ class Projects(Resource):
                 'required': True
             },
             {
+                'name': 'pathogen_id',
+                'description': 'Pathogen ID',
+                'in': 'formData',
+                'type': 'integer',
+                'required': True
+            },
+            {
                 'name': 'group',
                 'description': 'Project group',
                 'in': 'formData',
@@ -100,7 +160,7 @@ class Projects(Resource):
                 'description': 'Project description',
                 'in': 'formData',
                 'type': 'string',
-                'required': True
+                'required': False
             }
         ],
         'responses': {
@@ -113,23 +173,32 @@ class Projects(Resource):
         parser = reqparse.RequestParser()
         parser.add_argument('title', required=True)
         parser.add_argument('pid', required=True)
+        parser.add_argument('pathogen_id', required=True)
         parser.add_argument('group', required=True)
-        parser.add_argument('description', required=True)
+        parser.add_argument('description', required=False)
         args = parser.parse_args()
 
-       
-        new_project = Project(
-            title=args['title'],
-            pid=args['pid'],
-            group=args['group'],
-            description=args['description'],
-            owner_id=get_jwt_identity()
-        )
 
-        db.session.add(new_project)
-        db.session.commit()
+        claims = get_jwt()
+        if app.config['PROJECT_GROUP'] in claims['context']['user']['groups']:
 
-        return new_project.as_dict(), 201
+            new_project = Project(
+                title=args['title'],
+                pid=args['pid'],
+                pathogen_id=args['pathogen_id'],
+                group=args['group'],
+                description=args['description'],
+                owner_id=get_jwt_identity()
+            )
+
+            db.session.add(new_project)
+            db.session.commit()
+
+            return new_project.as_dict(), 201
+    
+        else:
+            return {'message': 'You do not have the required permissions to create a project'}, 401
+        
 
     @jwt_required()
     @swagger.doc({
@@ -148,21 +217,28 @@ class Projects(Resource):
                 'description': 'Project title',
                 'in': 'formData',
                 'type': 'string',
-                'required': False
+                'required': True
             },
             {
                 'name': 'pid',
                 'description': 'Project ID',
                 'in': 'formData',
                 'type': 'string',
-                'required': False
+                'required': True
+            },
+            {
+                'name': 'pathogen_id',
+                'description': 'Pathogen ID',
+                'in': 'formData',
+                'type': 'integer',
+                'required': True
             },
             {
                 'name': 'group',
                 'description': 'Project group',
                 'in': 'formData',
                 'type': 'string',
-                'required': False
+                'required': True
             },
             {
                 'name': 'description',
@@ -183,11 +259,12 @@ class Projects(Resource):
         parser.add_argument('id', type=int, required=True)
         parser.add_argument('title')
         parser.add_argument('pid')
+        parser.add_argument('pathogen_id')
         parser.add_argument('group')
         parser.add_argument('description')
         args = parser.parse_args()
 
-        project = Project.query.get(args['id'])
+        project = db.session.get(Project, args['id'])
         if not project:
             return {'message': 'Project not found'}, 404
 
@@ -195,12 +272,18 @@ class Projects(Resource):
             project.title = args['title']
         if args['pid']:
             project.pid = args['pid']
+        if args['pathogen_id']:
+            project.pathogen_id = args['pathogen_id']
         if args['group']:
             project.group = args['group']
         if args['description']:
             project.description = args['description']
 
-        db.session.commit()
+        claims = get_jwt()
+        if app.config['PROJECT_GROUP'] in claims['context']['user']['groups']:
+            db.session.commit()
+        else:
+            return {'message': 'You do not have the required permissions to update this project'}, 401
 
         return project.as_dict(), 200
     
@@ -228,14 +311,19 @@ class Projects(Resource):
         parser.add_argument('id', type=int, required=True)
         args = parser.parse_args()
 
-        project = Project.query.get(args['id'])
+        project = db.session.get(Project, args['id'])
         if not project:
             return {'message': 'Project not found'}, 404
+        
+        claims = get_jwt()
+        if app.config['PROJECT_GROUP'] in claims['context']['user']['groups']:
+            db.session.delete(project)
+            db.session.commit()
+            return {'message': 'Project deleted'}, 204
+        else:
+            return {'message': 'You do not have the required permissions to delete this project'}, 401
 
-        db.session.delete(project)
-        db.session.commit()
-
-        return {}, 204
+        
     
 class Pathogens(Resource):
 
@@ -263,14 +351,166 @@ class Pathogens(Resource):
             pathogen_list.append(pathogen_data)
         return pathogen_list
     
+
+    @jwt_required()
+    @swagger.doc({
+        'tags': ['pathogens'],
+        'description': 'Create a new pathogen',
+        'parameters': [
+            {
+                'name': 'common_name',
+                'description': 'Common name',
+                'in': 'formData',
+                'type': 'string',
+                'required': True
+            },
+            {
+                'name': 'scientific_name',
+                'description': 'Scientific name',
+                'in': 'formData',
+                'type': 'string',
+                'required': True
+            },
+            {
+                'name': 'schema',
+                'description': 'Pathogen schema',
+                'in': 'formData',
+                'type': 'string',
+                'required': False
+            }
+        ],
+        'responses': {
+            '201': {
+                'description': 'Pathogen created'
+            }
+        }
+    })
     def post(self):
-        pass
+        parser = reqparse.RequestParser()
+        parser.add_argument('common_name', required=True)
+        parser.add_argument('scientific_name', required=True)
+        parser.add_argument('schema')
+        args = parser.parse_args()
+
+        claims = get_jwt()
+        if app.config['PATHOGEN_GROUP'] in claims['context']['user']['groups']:
+            new_pathogen = Pathogen(
+                common_name=args['common_name'],
+                scientific_name=args['scientific_name'],
+                schema=args['schema']
+            )
+
+            db.session.add(new_pathogen)
+            db.session.commit()
+
+            return new_pathogen.as_dict(), 201
+        else:
+            return {'message': 'You do not have the required permissions to create a pathogen'}, 401
     
+
+    @jwt_required()
+    @swagger.doc({
+        'tags': ['pathogens'],
+        'description': 'Update a pathogen',
+        'parameters': [
+            {
+                'name': 'id',
+                'description': 'Pathogen ID',
+                'in': 'formData',
+                'type': 'integer',
+                'required': True
+            },
+            {
+                'name': 'common_name',
+                'description': 'Common name',
+                'in': 'formData',
+                'type': 'string',
+                'required': False
+            },
+            {
+                'name': 'scientific_name',
+                'description': 'Scientific name',
+                'in': 'formData',
+                'type': 'string',
+                'required': False
+            },
+            {
+                'name': 'schema',
+                'description': 'Pathogen schema',
+                'in': 'formData',
+                'type': 'string',
+                'required': False
+            }
+        ],
+        'responses': {
+            '200': {
+                'description': 'Pathogen updated'
+            }
+        }
+    })
     def put(self):
-        pass
+        parser = reqparse.RequestParser()
+        parser.add_argument('id', type=int, required=True)
+        parser.add_argument('common_name')
+        parser.add_argument('scientific_name')
+        parser.add_argument('schema')
+        args = parser.parse_args()
+
+        pathogen = db.session.get(Pathogen, args['id'])
+        if not pathogen:
+            return {'message': 'Pathogen not found'}, 404
+
+        if args['common_name']:
+            pathogen.common_name = args['common_name']
+        if args['scientific_name']:
+            pathogen.scientific_name = args['scientific_name']
+        if args['schema']:
+            pathogen.schema = args['schema']
+
+        claims = get_jwt()
+        if app.config['PATHOGEN_GROUP'] in claims['context']['user']['groups']:
+            db.session.commit()
+        else:
+            return {'message': 'You do not have the required permissions to update this pathogen'}, 401
+
+        return pathogen.as_dict(), 200
     
+    @jwt_required()
+    @swagger.doc({
+        'tags': ['pathogens'],
+        'description': 'Delete a pathogen',
+        'parameters': [
+            {
+                'name': 'id',
+                'description': 'Pathogen ID',
+                'in': 'formData',
+                'type': 'integer',
+                'required': True
+            }
+        ],
+        'responses': {
+            '204': {
+                'description': 'Pathogen deleted'
+            }
+        }
+    })
     def delete(self):
-        pass
+        parser = reqparse.RequestParser()
+        parser.add_argument('id', type=int, required=True)
+        args = parser.parse_args()
+
+        pathogen = db.session.get(Pathogen, args['id'])
+        if not pathogen:
+            return {'message': 'Pathogen not found'}, 404
+        
+        claims = get_jwt()
+        if app.config['PATHOGEN_GROUP'] in claims['context']['user']['groups']:
+            db.session.delete(pathogen)
+            db.session.commit()
+            return {'message': 'Pathogen deleted'}, 204
+        else:
+            return {'message': 'You do not have the required permissions to delete this pathogen'}, 401
+
 
 class Studies(Resource):
 
@@ -294,48 +534,12 @@ class Studies(Resource):
     
     def delete(self):
         pass
-    
-class GenerateToken(Resource):
-    def get(self):
-        user_payload = {
-            "iat": datetime.datetime.utcnow(),
-            "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=1),
-            "sub": "3423616",
-            "iss": "ego",
-            "jti": "2323232323",
-            "context": {
-                "scope": [
-                    "VIRUS-SEQ.WRITE",
-                    "VIRUS-SEQ.READ"
-                ],
-                "user": {
-                    "email": "user@domain.com",
-                    "status": "APPROVED",
-                    "firstName": "Di",
-                    "lastName": "MMM",
-                    "createdAt": 1715867872506,
-                    "lastLogin": 1717576566031,
-                    "preferredLanguage": None,
-                    "providerType": "KEYCLOAK",
-                    "providerSubjectId": "eee",
-                    "type": "ADMIN",
-                    "groups": [
-                        "1111"
-                    ]
-                }
-            },
-            "aud": []
-        }
-        
-        token = jwt.encode(user_payload, app.config['JWT_SECRET_KEY'], algorithm='HS256')
-        return jsonify({'token': token})
 
 
 api.add_resource(User, '/api/')
 api.add_resource(Projects, '/api/projects/')
 api.add_resource(Pathogens, '/api/pathogens/')
 api.add_resource(Studies, '/api/studies/')
-api.add_resource(GenerateToken, '/api/generate-token')
 
 
 if __name__ == '__main__':
